comments for main method, execution stage and alu logic

+++ src/structures.h
@@ -1,5 +1,9 @@
-
+#include <stdint.h>
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
 #include "processor.h"
+// added the imports 34an kanet 3amla error
 
 // Pipeline registers between stages
 typedef struct {
@@ -21,7 +25,7 @@ typedef struct {
 } ID_EX_Reg;
 
 typedef struct {
-    int      dest_reg;      // register to write back (-1 = no writeback)
+    int      dest_reg;      // register to write back (-1 = no writeback) rakamo ya farah focussss
     int8_t   result;        // ALU or memory result
     uint16_t new_pc;        // updated PC if branch/jump taken
     FlushType flush;
@@ -33,11 +37,11 @@ typedef struct {
     int8_t   regs[64];          // R0–R63, 8-bit general purpose
     uint16_t pc;                // 16-bit program counter
     uint8_t  sreg;              // bits: 0=Z,1=S,2=N,3=V,4=C (bits7:5 = 0)
-    uint8_t  instr_mem[1024];   // 16-bit words stored as byte pairs
+    uint16_t  instr_mem[1024];   // 16-bit words stored as byte pairs (kanet ma7tota 8 8ayrtha l 16)
     uint8_t  data_mem[2048];    // 8-bit per address
-    int      instr_count;       // number of instructions loaded
+    int      instr_count;       // number of instructions loaded (it was stated to be 12 ??? but at the same time depends on the program)
     int      clock_cycle;
-    IF_ID_Reg  if_id;
-    ID_EX_Reg  id_ex;
-    EX_Result  ex_out;
+    IF_ID_Reg  if_id; // inst. fetch
+    ID_EX_Reg  id_ex; // inst. Decode
+    EX_Result  ex_out; // exec. result
 } ProcessorState;
